app/requirements.txt
PyGithub
plotly
python-dotenv
streamlit
streamlit-option-menu
openpyxl
kaleido
# Testing dependencies
pytest>=6.2.5
pytest-cov>=2.12.1
freezegun>=1.2.0